requirements.txt
# =============================================================================
# Requirements — Conser-vision: Wildlife Image Classification
# Competition: https://www.drivendata.org/competitions/87/
# Generated: 2026-03-29
# Python: 3.10+
# =============================================================================

# --- Core ---
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.5.0
Pillow==10.3.0

# --- Deep Learning ---
torch==2.3.0
torchvision==0.18.0
timm==1.0.3                   # EfficientNet, ConvNeXt, ViT pretrained models

# --- Augmentation ---
albumentations==1.4.7         # ricco di trasformazioni image-specific

# --- Training utils ---
tqdm==4.66.4
pyyaml==6.0.1
python-dotenv==1.0.1

# --- Metrics & Evaluation ---
# scikit-learn già include log_loss

# --- Visualisation ---
matplotlib==3.9.0
seaborn==0.13.2
opencv-python-headless==4.10.0.82

# --- Experiment tracking (decommentare se usato) ---
# wandb==0.17.0

# --- Utilities ---
joblib==1.4.2
psutil==5.9.8

# --- Download data ---
# drivendata-download  # installare manualmente: pip install drivendata-download

# --- Dev ---
# jupyter
# ipykernel
# ruff
